added random and auto-deletion

+++ Bot.py
@@ -1,15 +1,18 @@
-import asyncio
 import os
+import random
+import time
+import asyncio
 import subprocess
-import nest_asyncio
 import Credentials
-from datetime import datetime, timedelta
+import nest_asyncio
+
 from telegram.ext import Application
+from datetime import datetime, timedelta
 
 nest_asyncio.apply()
 
-
-STORIES_DIR = os.path.join(os.path.dirname(__file__), 'InstaStoryLoader/stories/'+Credentials.STORIES_PROVIDER_USERNAME)
+STORIES_DIR = os.path.join(os.path.dirname(__file__),
+                           'InstaStoryLoader/stories/' + Credentials.STORIES_PROVIDER_USERNAME)
 SENT_STORIES_FILE = 'sent_stories.txt'  # File to keep track of sent stories
 
 
@@ -35,15 +38,31 @@ def write_sent_stories(sent_stories):
 
 
 def clean_stories_folder():
+    current_time = time.time()
+    two_days = 172800
+
+    with open(SENT_STORIES_FILE, 'r') as file:
+        sent_stories = file.read().splitlines()
+
     for filename in os.listdir(STORIES_DIR):
         file_path = os.path.join(STORIES_DIR, filename)
-        os.remove(file_path)
-    open(SENT_STORIES_FILE, 'w').close()
+
+        modification_time = os.path.getmtime(file_path)
+
+        if current_time - modification_time > two_days:
+            os.remove(file_path)
+            if filename in sent_stories:
+                sent_stories.remove(filename)
+
+    with open(SENT_STORIES_FILE, 'w') as file:
+        for story in sent_stories:
+            file.write(story + '\n')
 
 
 async def download_stories():
+    random_username = [Credentials.INSTAGRAM_USERNAME, Credentials.INSTAGRAM_USERNAME1][random.randint(1,2)]
     command = ['python3', 'InstaStoryLoader/StoryLoader.py', '-u',
-               Credentials.INSTAGRAM_USERNAME, '-p', Credentials.INSTAGRAM_PASSWORD, '-d',
+               random_username, '-p', Credentials.INSTAGRAM_PASSWORD, '-d',
                Credentials.STORIES_PROVIDER_USERNAME, '--no-thumbs', '--taken-at']
     subprocess.run(command)
 
@@ -71,18 +90,29 @@ async def send_new_stories(context):
     write_sent_stories(sent_stories)
 
 
+async def remove_credentials_periodically():
+    credentials_file_path = os.path.join(os.path.dirname(__file__), 'InstaStoryLoader/credentials.json')
+    while True:
+        if os.path.exists(credentials_file_path):
+            os.remove(credentials_file_path)
+            print('[I] Credentials file removed.')
+        else:
+            print('[I] No credentials file found to remove.')
+        sleep_time = random.randint(13000, 15800)
+        await asyncio.sleep(sleep_time)
+
+
 async def periodic_task(context):
     last_cleanup = datetime.now()
     while True:
         await download_stories()
         await send_new_stories(context)
 
-        # Check if a day has passed to clean up the folder
-        if datetime.now() - last_cleanup >= timedelta(days=1):
+        if datetime.now() - last_cleanup >= timedelta(days=2):
             clean_stories_folder()
             last_cleanup = datetime.now()
-
-        await asyncio.sleep(3600)  # Wait for 1 hour
+        sleep_time = random.randint(3000, 4200)
+        await asyncio.sleep(sleep_time)
 
 
 async def main():
@@ -94,6 +124,7 @@ async def main():
         os.remove(feed_file_path)
     application = Application.builder().token(Credentials.BOT_KEY).build()
     asyncio.create_task(periodic_task(application))
+    asyncio.create_task(remove_credentials_periodically())
     await application.run_polling()
